Solve part two of day 24 in Rust

+++ Rust/day24/src/linalg.rs
@@ -1,4 +1,4 @@
-pub(crate) fn cross_product(a: [i64; 3], b: [i64; 3]) -> [i64; 3] {
+pub(crate) fn cross_product(a: [f64; 3], b: [f64; 3]) -> [f64; 3] {
     [
         a[1] * b[2] - a[2] * b[1],
         a[2] * b[0] - a[0] * b[2],
@@ -6,6 +6,6 @@ pub(crate) fn cross_product(a: [i64; 3], b: [i64; 3]) -> [i64; 3] {
     ]
 }
 
-pub(crate) fn dot_product(a: [i64; 3], b: [i64; 3]) -> i64 {
+pub(crate) fn dot_product(a: [f64; 3], b: [f64; 3]) -> f64 {
     a[0] * b[0] + a[1] * b[1] + a[2] * b[2]
 }
